--- src/modules/positions/actions/clear-account-trades.js
@@ -1,5 +1,0 @@
-export const CLEAR_ACCOUNT_TRADES = "CLEAR_ACCOUNT_TRADES";
-
-export function clearAccountTrades() {
-  return { type: CLEAR_ACCOUNT_TRADES };
-}
