+++ src/modules/positions/actions/load-account-trades.js
@@ -1,13 +1,17 @@
 import { parallel } from "async";
 import { augur } from "services/augurjs";
-import { clearAccountTrades } from "modules/positions/actions/clear-account-trades";
 import { addTradeTransactions } from "modules/transactions/actions/add-transactions";
 import { loadAccountPositions } from "modules/positions/actions/load-account-positions";
-import { loadAccountOrders } from "modules/orders/actions/load-account-orders";
+import {
+  loadAccountOrders,
+  loadAccountOrphanedOrders
+} from "modules/orders/actions/load-account-orders";
 import { loadMarketsInfoIfNotLoaded } from "modules/markets/actions/load-markets-info";
-import { updateAccountTradeData } from "modules/positions/actions/update-account-trades-data";
+import {
+  updateAccountTradeData,
+  clearAccountTrades
+} from "modules/positions/actions/update-account-trades-data";
 import logError from "utils/log-error";
-import { loadAccountOrphanedOrders } from "modules/orders/actions/load-account-orphaned-orders";
 
 export function loadAccountTrades(options, callback = logError) {
   return (dispatch, getState) => {
