+++ src/modules/positions/actions/update-account-trades-data.js
@@ -1,6 +1,10 @@
 export const UPDATE_ACCOUNT_TRADES_DATA = "UPDATE_ACCOUNT_TRADES_DATA";
 export const UPDATE_ACCOUNT_POSITIONS_DATA = "UPDATE_ACCOUNT_POSITIONS_DATA";
+export const CLEAR_ACCOUNT_TRADES = "CLEAR_ACCOUNT_TRADES";
 
+export function clearAccountTrades() {
+  return { type: CLEAR_ACCOUNT_TRADES };
+}
 export const updateAccountPositionsData = (data, marketId) => ({
   type: UPDATE_ACCOUNT_POSITIONS_DATA,
   data,
