+++ test/positions/reducers/account-trades-test.js
@@ -1,7 +1,9 @@
 import accountTrades from "modules/positions/reducers/account-trades";
 
-import { UPDATE_ACCOUNT_TRADES_DATA } from "modules/positions/actions/update-account-trades-data";
-import { CLEAR_ACCOUNT_TRADES } from "modules/positions/actions/clear-account-trades";
+import {
+  UPDATE_ACCOUNT_TRADES_DATA,
+  CLEAR_ACCOUNT_TRADES
+} from "modules/positions/actions/update-account-trades-data";
 import { CLEAR_LOGIN_ACCOUNT } from "modules/auth/actions/update-login-account";
 
 describe("modules/positions/reducers/account-trades.js", () => {
